refactor(#800): split commands/doctor.rs by moving tests to sibling
Move 2,996-line commands/doctor.rs into commands/doctor/ directory:

- doctor/mod.rs:   1,830 LOC (production code)
- doctor/tests.rs: 1,165 LOC (unit tests)

Both well below the 2,500 LOC limit. No behavior changes.

`cargo test -p budi-cli` green (281 passed).
